[ok] refactored program basic/tunserver to files; [fix] mo reader
