+++ .gitignore
@@ -1,3 +1,7 @@
+
+# the default directory for building
+build/
+
 # temporary var files, including created by running-in-place program
 # especially for developer tools, script
 vartmp/
