+++ src/libs1.hpp
@@ -1,6 +1,5 @@
 // Copyrighted (C) 2015-2016 Antinet.org team, see file LICENCE-by-Antinet.txt
 
-
 #pragma once
 #ifndef LIBS1_HPP
 #define LIBS1_HPP
@@ -8,6 +7,9 @@
 #include "libs0.hpp"
 
 #include "strings_utils.hpp"
+
+#include "utils/misc.hpp"
+#include "utils/check.hpp"
 
 #endif
